don't accidentally clobber an existing endpoint
